moved networking code from hwSetup to network to make files smaller/better to understand. Names of functions/variables not yet refactored
